python-seeder/requirements.txt
pandas==2.0.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0